requirements.txt
pytest==8.3.5
requests==2.32.3
Faker==24.8.0
allure-pytest==2.14.1
allure-python-commons==2.14.1
